Remove <algorithm>
Scaling test:

8 threads vs 1 thread: 1064 - 115 - 741  [0.747] 1920
ELO difference: 188.20 +/- 12.50

Testing conditions:

./cutechess-cli \
  -each proto=uci tc=8+0.08 option.Hash=32 \
  -variant 'fischerandom' \
  -openings file='./chess960.epd' format=epd -repeat \
  -engine cmd=$1 name="8 threads" option.Threads=8 \
  -engine cmd=$2 name="1 thread" \
  -pgnout $3 -recover \
  -concurrency 1 -rounds 1920 -ratinginterval 20 \
  -draw movenumber=40 movecount=8 score=20 \
  -resign movecount=3 score=500

CPU: i7-6700, 4 cores (8 threads)

+++ src/move.cc
@@ -145,7 +145,7 @@ int move_see(const Position *pos, move_t m)
     bitboard_t occ = pos_pieces(pos);
 
     // General case
-    int gain[32] = {see_value[pos->pieceOn[to]]};
+    int gain[32] = {see_value[(int)pos->pieceOn[to]]};
     int capture = pos->pieceOn[from];
     bb_clear(&occ, from);
